requirements.txt
# Python dependencies for venue scraping worker

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# LLM integration
groq>=0.4.0

# Database
psycopg2-binary>=2.9.9

# Task queue
celery>=5.3.0
redis>=5.0.0

# Web framework (optional, for health checks)
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0

